requirements.txt
certifi==2025.4.26
charset-normalizer==2.0.12
idna==3.10
python-dateutil==2.9.0.post0
requests==2.27.1
six==1.17.0
urllib3==1.26.20
